Work in progress MC/DC support in gcc/gcov
This is a *work in progress*, and I will have no regard for history of
this branch - as I find defects and make improvements I will happily
squash, so do not rely on it for history tracking. It is only published
for reference and mirroring.

Once it shapes up properly it will be proposed on the gcc mailing list.